requirements.txt
python-dotenv==1.0.0
openai==1.30.0
httpx==0.27.2
pyTelegramBotAPI==4.16.0
Flask==3.0.0
requests==2.31.0
